+++ LookupAnything/Framework/Fields/CharacterFriendshipField.cs
@@ -81,7 +81,7 @@ internal class CharacterFriendshipField : GenericField
 
         // get caption text
         string? caption = null;
-        if (this.Friendship.EmptyHearts == 0 && this.Friendship.LockedHearts > 0)
+        if (this.Friendship is { EmptyHearts: 0, LockedHearts: > 0 })
             caption = $"({I18n.Npc_Friendship_NeedBouquet()})";
         else
         {